requirements.txt
fastapi
SQLAlchemy
databases
uvicorn
pydantic
asyncpg
alembic
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]